Remove legacy modules, add stream router, fix warnings
Cleanup:
- Remove emu_stub module (replaced by interpreter/)
- Remove patterns.rs decoder (replaced by tablegen decoder)
- Rename tablegen_decoder.rs to decoder.rs
- Remove outdated tablegen-assessment.md

New:
- Add StreamRouter for global tile-to-tile data flow

Fixes:
- Fix 3 unused variable warnings (slot_layout, context, parser)
- Apply clippy auto-fixes (length checks, filter patterns)

All 433 tests passing.

+++ src/interpreter/execute/scalar.rs
@@ -73,7 +73,7 @@ impl ScalarAlu {
 
             Operation::ScalarShl => {
                 let (a, b) = Self::get_two_sources(op, ctx);
-                let shift = (b & 0x1F) as u32;
+                let shift = (b & 0x1F);
                 let result = a << shift;
                 Self::write_dest(op, ctx, result);
                 ctx.set_flags(Flags::from_result(result));
@@ -82,7 +82,7 @@ impl ScalarAlu {
 
             Operation::ScalarShr => {
                 let (a, b) = Self::get_two_sources(op, ctx);
-                let shift = (b & 0x1F) as u32;
+                let shift = (b & 0x1F);
                 let result = a >> shift;
                 Self::write_dest(op, ctx, result);
                 ctx.set_flags(Flags::from_result(result));
@@ -91,7 +91,7 @@ impl ScalarAlu {
 
             Operation::ScalarSra => {
                 let (a, b) = Self::get_two_sources(op, ctx);
-                let shift = (b & 0x1F) as u32;
+                let shift = (b & 0x1F);
                 let result = ((a as i32) >> shift) as u32;
                 Self::write_dest(op, ctx, result);
                 ctx.set_flags(Flags::from_result(result));
